BackEnd/requirements.txt
# BackEnd/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
motor>=3.1.1  # MongoDB async driver
python-jose>=3.3.0  # For JWT tokens
passlib>=1.7.4  # Password hashing
python-multipart>=0.0.6  # For handling form data (file uploads)
langchain>=0.0.267
pydantic>=1.10.7
python-dotenv>=1.0.0
transformers>=4.30.2
torch>=2.0.0
pymupdf>=1.22.3  # For PDF processing
python-docx>=0.8.11  # For DOCX processing
redis>=4.5.5
elasticsearch>=8.8.0